Improve consistency
