+++ react/features/base/styles/index.js
@@ -1,2 +1,2 @@
 export * from './components';
-export * from './createStyleSheet';
+export * from './functions';
